+++ src/main/java/insane96mcp/progressivebosses/event/EntityRegisterAttributes.java
@@ -9,6 +9,7 @@ import net.minecraftforge.event.entity.EntityAttributeCreationEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
+//TODO Move
 @Mod.EventBusSubscriber(modid = ProgressiveBosses.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
 public class EntityRegisterAttributes {
 	@SubscribeEvent