+++ src/main/java/insane96mcp/progressivebosses/event/PlayerClone.java
@@ -7,6 +7,7 @@ import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
+//TODO Move
 @Mod.EventBusSubscriber(modid = ProgressiveBosses.MOD_ID)
 public class PlayerClone {