+++ src/main/java/insane96mcp/progressivebosses/mixin/EndDragonFightMixin.java
@@ -12,7 +12,6 @@ import net.minecraft.world.level.dimension.end.EndDragonFight;
 import net.minecraft.world.level.levelgen.feature.SpikeFeature;
 import net.minecraft.world.phys.AABB;
 import org.objectweb.asm.Opcodes;
-import org.spongepowered.asm.mixin.Debug;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -27,7 +26,6 @@ import java.util.List;
 import java.util.Optional;
 
 @Mixin(EndDragonFight.class)
-@Debug(export = true)
 public class EndDragonFightMixin {
 
 	@Shadow @Final private ServerLevel level;